setup(security)Added security check [bind_param, md5 ...]

+++ src/manager/UserManager.php
@@ -24,27 +24,30 @@ class UserManager extends Manager
      * @param $userId
      * @return array|mixed
      */
-    public function getUser ( $userId)
+    public function getUser ($userId)
     {
-        return $this->prepareObject('SELECT * FROM user WHERE id=' . $userId, User::class, false);
+        return $this->prepareObject('SELECT * FROM user WHERE id= :userId', User::class, false,
+            [':userId' => $userId]);
     }
 
     /**
      * @param $postId
      * @return array|mixed
      */
-    public function getPostUser( $postId)
+    public function getPostUser($postId)
     {
-        return $this->prepareObject('SELECT * FROM user WHERE id =' . $postId, User::class, false);
+        return $this->prepareObject('SELECT * FROM user WHERE id = :postId', User::class, false,
+            [':postId' => $postId]);
     }
 
     /**
      * @param $commentId
      * @return array|mixed
      */
-    public function getCommentUser( $commentId)
+    public function getCommentUser($commentId)
     {
-        return $this->prepareObject('SELECT * FROM user WHERE id =' . $commentId, User::class, false);
+        return $this->prepareObject('SELECT * FROM user WHERE id = :commentId', User::class,
+        false, [':commentId' => $commentId]);
     }
 
     /**
@@ -52,9 +55,14 @@ class UserManager extends Manager
      * @param $password
      * @param $email
      */
-    public function addUser( $username, $password, $email)
+    public function addUser($username, $password, $email)
     {
-        $this->prepareStmt('INSERT INTO user (username, password, email) VALUES ("'. $username . '", "'. $password . '", "'. $email . '")');
+        $securedPass = md5($password);
+        $this->prepareStmt(
+        'INSERT INTO user (username, password, email)
+        VALUES (:username, :securedPass, :email)',
+        [':username' => $username, ':securedPass' => $securedPass, ':email' => $email]
+        );
     }
 
     /**
@@ -63,9 +71,13 @@ class UserManager extends Manager
      * @param $email
      * @return array|mixed
      */
-    public function modifyUser( $userId, $username, $email)
+    public function modifyUser($userId, $username, $email)
     {
-        $this->prepareStmt('UPDATE user SET username = "' . $username . '", email = "' . $email . '" WHERE id=' . $userId);
+        $this->prepareStmt(
+        'UPDATE user SET username = :username, email = :email
+        WHERE id= :userId',
+        [':username' => $username, ':email' => $email, ':userId' => $userId]
+        );
         return $this->getUser($userId);
     }
 
@@ -74,19 +86,19 @@ class UserManager extends Manager
      * @param $password
      * @return array|mixed
      */
-    public function modifyUserPass( $userId, $password)
+    public function modifyUserPass($userId, $password)
     {
-        $this->prepareStmt('UPDATE user SET password = "' . $password . '" WHERE id=' . $userId);
+        $this->prepareStmt('UPDATE user SET password = :password WHERE id= :userId', [':userId' => $userId, ':password' => $password]);
         return $this->getUser($userId);
     }
 
     /**
      * @param $userId
      */
-    public function deleteUser( $userId)
+    public function deleteUser($userId)
     {
         try{
-            return $this->prepareStmt('DELETE FROM user WHERE id=' . $userId);
+            return $this->prepareStmt('DELETE FROM user WHERE id= :userId', [':userId' => $userId]);
         }
         catch (\PDOException $e) {
             $erreur = explode(':', $e->getMessage());
@@ -100,26 +112,30 @@ class UserManager extends Manager
      * @param $username
      * @param $password
      */
-    public function checkUser( $username, $password)
+    public function checkUser($username, $password)
     {
-        $_SESSION['user'] = $this->prepareObject('SELECT * FROM user where username = "'. $username . '" && password = "'. $password . '"', User::class, false );
+        $_SESSION['user'] = $this->prepareObject(
+        'SELECT * FROM user where username = :username AND password = :password', User::class,
+        false, [':username' => $username, ':password' => $password]);
     }
 
     /**
      * @param $email
      */
-    public function checkUserEmail( $email)
+    public function checkUserEmail($email)
     {
-        $_SESSION['email'] = $this->prepareObject('SELECT * FROM user where email = "'. $email . '"', User::class, false );
+        $_SESSION['email'] = $this->prepareObject(
+        'SELECT * FROM user where email = :email', User::class, false, [':email' => $email]);
     }
 
-    public function newPass( $userId, $newpassword )
+    public function newPass($userId, $newpassword )
     {
-        $this->prepareStmt('UPDATE user SET password = "' . $newpassword . '", newpass = "1" WHERE id=' . $userId);
+        $this->prepareStmt('UPDATE user SET password = :newpassword, newpass = "1"
+        WHERE id= :userId', [':newpassword' => $newpassword, ':userId' => $userId]);
     }
 
-    public function refreshNewpass( $userId )
+    public function refreshNewpass($userId)
     {
-        $this->prepareStmt('UPDATE user SET newpass = "0" WHERE id=' . $userId);
+        $this->prepareStmt('UPDATE user SET newpass = "0" WHERE id= :userId', [':userId' => $userId]);
     }
 }
